+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - add option to open message html in browser
 - encryption info for groups
+- allow sending of ".webp" stickers
 
 ## Fixed
 - hide show encryption info for saved messages (resulted in error)